Compare counts and labels of a list of dataframes

Compare counts and labels of a list of dataframes

cmp_all(
  l,
  id = "id",
  include_ids = FALSE,
  spec_diffs = c("nv", "cv", "vallab", "varlab"),
  include_diffs = TRUE,
  col_groups = c("spec", "index")
)

Arguments

l
List of dataframes.

id
name of the key variable in the dataframes.

include_ids
Logical denoting whether a list column ids should be included in the results. The ids in each list show at which values of id the variable var contains the value val1, val2, ...

spec_diffs
Character vector (defaults to c("ex", "varlab", "val", "vallab")) specifying the attributes that should be compared in the resulting dataframe. If include_diffs = TRUE, this adds columns prefixed by spec_diffs, and each suffixed by "_diff". Additionally a column any_diff is added denoting if any of the former is TRUE.

include_diffs
Logical denoting whether in the resulting dataframe, logical columns suffixes "_diff" should be included, indicating if there are differences between the dataframes in l regarding the specification types given in spec_diffs.

col_groups
String specifying the order of columns in the resulting dataframe. "index" groups columns by indices. "spec" groups columns by the specification types c("ex", "varlab", "val", "vallab")

Value

Dataframe consisting of columns var, val1, val2, ..., vallab1, vallab2, ..., varlab1, varlab2, ..., ex1, ex2, ..., and n, containing a comparison of the counts of variable values (and their respective value labels) of the dataframes in long format. The indices denote the dataframes in l. vals_diff & vallabs_diff are logical columns indicating if all values, variable / value labels are equal.

Examples

# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) %>%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
df2[1, "Species"] <- 2
# modify the value label of "setosa"
df2$Species <- haven::labelled(df2$Species,
                               labels = c(setosa_mod = 1, versicolor = 2, virginica = 3),
                               label = "Species")

# compare the dataframes counts:
cmp_all(list(df1, df2))

#> No variable in the data.frame has a variable label

#> # A tibble: 127 × 15
#>    var              n   nv1   nv2 nv_diff cv1   cv2   cv_diff vallab1 vallab2
#>    <chr>        <int> <dbl> <dbl> <lgl>   <chr> <chr> <lgl>   <chr>   <chr>  
#>  1 Sepal.Length     1   4.3   4.3 FALSE   NA    NA    FALSE   NA      NA     
#>  2 Sepal.Length     3   4.4   4.4 FALSE   NA    NA    FALSE   NA      NA     
#>  3 Sepal.Length     1   4.5   4.5 FALSE   NA    NA    FALSE   NA      NA     
#>  4 Sepal.Length     4   4.6   4.6 FALSE   NA    NA    FALSE   NA      NA     
#>  5 Sepal.Length     2   4.7   4.7 FALSE   NA    NA    FALSE   NA      NA     
#>  6 Sepal.Length     5   4.8   4.8 FALSE   NA    NA    FALSE   NA      NA     
#>  7 Sepal.Length     6   4.9   4.9 FALSE   NA    NA    FALSE   NA      NA     
#>  8 Sepal.Length    10   5     5   FALSE   NA    NA    FALSE   NA      NA     
#>  9 Sepal.Length     9   5.1   5.1 FALSE   NA    NA    FALSE   NA      NA     
#> 10 Sepal.Length     4   5.2   5.2 FALSE   NA    NA    FALSE   NA      NA     
#> # … with 117 more rows, and 5 more variables: vallab_diff <lgl>, varlab1 <chr>,
#> #   varlab2 <chr>, varlab_diff <lgl>, any_diff <lgl>

# compare the dataframes and only show the counts where values have changed:
cmp_all(list(df1, df2)) %>% dplyr::filter(nv1 != nv2)

#> No variable in the data.frame has a variable label

#> # A tibble: 1 × 15
#>   var         n   nv1   nv2 nv_diff cv1   cv2   cv_diff vallab1 vallab2   
#>   <chr>   <int> <dbl> <dbl> <lgl>   <chr> <chr> <lgl>   <chr>   <chr>     
#> 1 Species     1     1     2 TRUE    NA    NA    FALSE   setosa  versicolor
#> # … with 5 more variables: vallab_diff <lgl>, varlab1 <chr>, varlab2 <chr>,
#> #   varlab_diff <lgl>, any_diff <lgl>

# This results in the same rows:
cmp_all(list(df1, df2), spec_diffs = "nv") %>% dplyr::filter(any_diff)

#> No variable in the data.frame has a variable label

#> # A tibble: 1 × 6
#>   var         n   nv1   nv2 nv_diff any_diff
#>   <chr>   <int> <dbl> <dbl> <lgl>   <lgl>   
#> 1 Species     1     1     2 TRUE    TRUE    

# Or alternatively:
cmp_all(list(df1, df2)) %>% dplyr::filter(nv_diff)

#> No variable in the data.frame has a variable label

#> # A tibble: 1 × 15
#>   var         n   nv1   nv2 nv_diff cv1   cv2   cv_diff vallab1 vallab2   
#>   <chr>   <int> <dbl> <dbl> <lgl>   <chr> <chr> <lgl>   <chr>   <chr>     
#> 1 Species     1     1     2 TRUE    NA    NA    FALSE   setosa  versicolor
#> # … with 5 more variables: vallab_diff <lgl>, varlab1 <chr>, varlab2 <chr>,
#> #   varlab_diff <lgl>, any_diff <lgl>

# compare the dataframes and only show the counts where value labels have changed:
cmp_all(list(df1, df2)) %>% dplyr::filter(vallab1 != vallab2)

#> No variable in the data.frame has a variable label

#> # A tibble: 2 × 15
#>   var         n   nv1   nv2 nv_diff cv1   cv2   cv_diff vallab1 vallab2   
#>   <chr>   <int> <dbl> <dbl> <lgl>   <chr> <chr> <lgl>   <chr>   <chr>     
#> 1 Species    49     1     1 FALSE   NA    NA    FALSE   setosa  setosa_mod
#> 2 Species     1     1     2 TRUE    NA    NA    FALSE   setosa  versicolor
#> # … with 5 more variables: vallab_diff <lgl>, varlab1 <chr>, varlab2 <chr>,
#> #   varlab_diff <lgl>, any_diff <lgl>

# Create another modified copy
df3 <- df2
df3[2, "Species"] <- 3
# modify the value label of "versicolor"
df2$Species <- haven::labelled(df2$Species,
                               labels = c(setosa = 1, versicolor_mod = 2, virginica = 3),
                               label = "Species_mod")

# compare the dataframes counts:
l <- list(df1, df2, df3)
cmp <- cmp_all(l)

#> No variable in the data.frame has a variable label

cmp

#> # A tibble: 128 × 19
#>    var              n   nv1   nv2   nv3 nv_diff cv1   cv2   cv3   cv_diff vallab1
#>    <chr>        <int> <dbl> <dbl> <dbl> <lgl>   <chr> <chr> <chr> <lgl>   <chr>  
#>  1 Sepal.Length     1   4.3   4.3   4.3 FALSE   NA    NA    NA    FALSE   NA     
#>  2 Sepal.Length     3   4.4   4.4   4.4 FALSE   NA    NA    NA    FALSE   NA     
#>  3 Sepal.Length     1   4.5   4.5   4.5 FALSE   NA    NA    NA    FALSE   NA     
#>  4 Sepal.Length     4   4.6   4.6   4.6 FALSE   NA    NA    NA    FALSE   NA     
#>  5 Sepal.Length     2   4.7   4.7   4.7 FALSE   NA    NA    NA    FALSE   NA     
#>  6 Sepal.Length     5   4.8   4.8   4.8 FALSE   NA    NA    NA    FALSE   NA     
#>  7 Sepal.Length     6   4.9   4.9   4.9 FALSE   NA    NA    NA    FALSE   NA     
#>  8 Sepal.Length    10   5     5     5   FALSE   NA    NA    NA    FALSE   NA     
#>  9 Sepal.Length     9   5.1   5.1   5.1 FALSE   NA    NA    NA    FALSE   NA     
#> 10 Sepal.Length     4   5.2   5.2   5.2 FALSE   NA    NA    NA    FALSE   NA     
#> # … with 118 more rows, and 8 more variables: vallab2 <chr>, vallab3 <chr>,
#> #   vallab_diff <lgl>, varlab1 <chr>, varlab2 <chr>, varlab3 <chr>,
#> #   varlab_diff <lgl>, any_diff <lgl>

# compare the dataframes and only show the counts where values have changed:
cmp %>% dplyr::filter(nv_diff)

#> # A tibble: 2 × 19
#>   var         n   nv1   nv2   nv3 nv_diff cv1   cv2   cv3   cv_diff vallab1
#>   <chr>   <int> <dbl> <dbl> <dbl> <lgl>   <chr> <chr> <chr> <lgl>   <chr>  
#> 1 Species     1     1     1     3 TRUE    NA    NA    NA    FALSE   setosa 
#> 2 Species     1     1     2     2 TRUE    NA    NA    NA    FALSE   setosa 
#> # … with 8 more variables: vallab2 <chr>, vallab3 <chr>, vallab_diff <lgl>,
#> #   varlab1 <chr>, varlab2 <chr>, varlab3 <chr>, varlab_diff <lgl>,
#> #   any_diff <lgl>

# Show where either values or value labels differ:
cmp %>% dplyr::filter(nv_diff | vallab_diff)

#> # A tibble: 4 × 19
#>   var         n   nv1   nv2   nv3 nv_diff cv1   cv2   cv3   cv_diff vallab1   
#>   <chr>   <int> <dbl> <dbl> <dbl> <lgl>   <chr> <chr> <chr> <lgl>   <chr>     
#> 1 Species    48     1     1     1 FALSE   NA    NA    NA    FALSE   setosa    
#> 2 Species     1     1     1     3 TRUE    NA    NA    NA    FALSE   setosa    
#> 3 Species     1     1     2     2 TRUE    NA    NA    NA    FALSE   setosa    
#> 4 Species    50     2     2     2 FALSE   NA    NA    NA    FALSE   versicolor
#> # … with 8 more variables: vallab2 <chr>, vallab3 <chr>, vallab_diff <lgl>,
#> #   varlab1 <chr>, varlab2 <chr>, varlab3 <chr>, varlab_diff <lgl>,
#> #   any_diff <lgl>