Compare counts of a list of dataframes

Compare counts of a list of dataframes

cmp_cts(l, id = "id", include_ids = FALSE)

Arguments

l
List of dataframes.

id
name of the key variable in the dataframes.

include_ids
Logical denoting whether a list column of the ids should be included in the results. The ids in the list show at which values of id the variable var contains the numeric value nv1 or character (string) value cv1.

Value

Dataframe consisting of columns var, val, ..., vallab1, vallab2, ..., & n, containing a comparison of the counts of variable values (and their respective value labels) of the two dataframes in long format. nv_differ or cv_differ & vallabs_differ are logical columns indicating if all values / value labels are equal.

Examples

# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df1 <- haven::read_sav(path) %>%
  # add id column
  tibble::rownames_to_column("id")

# create a modified copy:
df2 <- df1
df2[1, "Species"] <- 2

# compare the dataframes counts:
cmp_cts(list(df1, df2))

#> # A tibble: 127 × 6
#>    var            nv1 cv1     nv2 cv2       n
#>    <chr>        <dbl> <chr> <dbl> <chr> <int>
#>  1 Sepal.Length   4.3 NA      4.3 NA        1
#>  2 Sepal.Length   4.4 NA      4.4 NA        3
#>  3 Sepal.Length   4.5 NA      4.5 NA        1
#>  4 Sepal.Length   4.6 NA      4.6 NA        4
#>  5 Sepal.Length   4.7 NA      4.7 NA        2
#>  6 Sepal.Length   4.8 NA      4.8 NA        5
#>  7 Sepal.Length   4.9 NA      4.9 NA        6
#>  8 Sepal.Length   5   NA      5   NA       10
#>  9 Sepal.Length   5.1 NA      5.1 NA        9
#> 10 Sepal.Length   5.2 NA      5.2 NA        4
#> # … with 117 more rows

# compare the dataframes and only show the counts where values have changed:
cmp_cts(list(df1, df2)) %>% dplyr::filter(nv1 != nv2)

#> # A tibble: 1 × 6
#>   var       nv1 cv1     nv2 cv2       n
#>   <chr>   <dbl> <chr> <dbl> <chr> <int>
#> 1 Species     1 NA        2 NA        1

# Create another modified copy
df3 <- df2
df3[2, "Species"] <- 3

# compare the dataframes counts:
l <- list(df1, df2, df3)
cmp <- cmp_cts(l)
cmp

#> # A tibble: 128 × 8
#>    var            nv1 cv1     nv2 cv2     nv3 cv3       n
#>    <chr>        <dbl> <chr> <dbl> <chr> <dbl> <chr> <int>
#>  1 Sepal.Length   4.3 NA      4.3 NA      4.3 NA        1
#>  2 Sepal.Length   4.4 NA      4.4 NA      4.4 NA        3
#>  3 Sepal.Length   4.5 NA      4.5 NA      4.5 NA        1
#>  4 Sepal.Length   4.6 NA      4.6 NA      4.6 NA        4
#>  5 Sepal.Length   4.7 NA      4.7 NA      4.7 NA        2
#>  6 Sepal.Length   4.8 NA      4.8 NA      4.8 NA        5
#>  7 Sepal.Length   4.9 NA      4.9 NA      4.9 NA        6
#>  8 Sepal.Length   5   NA      5   NA      5   NA       10
#>  9 Sepal.Length   5.1 NA      5.1 NA      5.1 NA        9
#> 10 Sepal.Length   5.2 NA      5.2 NA      5.2 NA        4
#> # … with 118 more rows

# compare the dataframes and only show the counts where values have changed:
cmp %>% dplyr::filter(!(nv1 == nv2 & nv2 == nv3))

#> # A tibble: 2 × 8
#>   var       nv1 cv1     nv2 cv2     nv3 cv3       n
#>   <chr>   <dbl> <chr> <dbl> <chr> <dbl> <chr> <int>
#> 1 Species     1 NA        1 NA        3 NA        1
#> 2 Species     1 NA        2 NA        2 NA        1