Compare the value labels between a list of dataframes

Compare the value labels between a list of dataframes

cmp_vallabs(l)

Arguments

l
A list of dataframes.

Value

A dataframe comparing the vallabs for all labelled values of the variables in the dataframes. The logical column vallab_diff indicates, if there are differences.

Examples

df <- data.frame(fbnr = 1:10,
sex = haven::labelled(c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                      label = "sex",
                      labels = c(MALES = 1, FEMALES = 2)),
                      age = c(24, 23, 23, 41, 23, 39, 30, 18, 31, 48))
df2 <- data.frame(fbnr = 1:10,
                  sex = haven::labelled(c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                                        label = "sex",
                                        labels = c(m = 0, f = 1)),
                  age = c(24, 23, 23, 41, 23, 39, 30, 18, 31, 48))
cmp_vallabs(list(df, df2))

#> # A tibble: 3 × 6
#>   var      nv vallab1 cv    vallab2 vallab_diff
#>   <chr> <dbl> <chr>   <chr> <chr>   <lgl>      
#> 1 sex       1 MALES   NA    f       TRUE       
#> 2 sex       2 FEMALES NA    NA      TRUE       
#> 3 sex       0 NA      NA    m       TRUE       

# The values are ordered by appearance in the dataframes. Compare:
cmp_vallabs(list(df2, df))

#> # A tibble: 3 × 6
#>   var      nv vallab1 cv    vallab2 vallab_diff
#>   <chr> <dbl> <chr>   <chr> <chr>   <lgl>      
#> 1 sex       0 m       NA    NA      TRUE       
#> 2 sex       1 f       NA    MALES   TRUE       
#> 3 sex       2 NA      NA    FEMALES TRUE