Create long format of a dataframe, keeping the id column.

Create long format of a dataframe, keeping the id column.

longen(df, id = "id")

Arguments

df

Dataframe with a key variable (id).

id

name of the key variable in the dataframe.

Value

Dataframe consisting of 3 columns id, var & val, containing the dataframe in long format (based on dplyr::gather).

Examples

# load spss data
path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path) %>%
  # add id column
  tibble::rownames_to_column("id")
df %>% longen()
#> # A tibble: 750 × 4
#>    id    var             nv cv   
#>    <chr> <chr>        <dbl> <chr>
#>  1 1     Sepal.Length   5.1 NA   
#>  2 10    Sepal.Length   4.9 NA   
#>  3 100   Sepal.Length   5.7 NA   
#>  4 101   Sepal.Length   6.3 NA   
#>  5 102   Sepal.Length   5.8 NA   
#>  6 103   Sepal.Length   7.1 NA   
#>  7 104   Sepal.Length   6.3 NA   
#>  8 105   Sepal.Length   6.5 NA   
#>  9 106   Sepal.Length   7.6 NA   
#> 10 107   Sepal.Length   4.9 NA   
#> # … with 740 more rows