Add a variable label and/or value labels (or none) to a variable.

Add a variable label and/or value labels (or none) to a variable.

set_labs(x, vallabs, varlab)

Arguments

x

A vector.

vallabs

Named vector of value labels.

varlab

A string containing the variable label.

Value

The function returns the variable x with variable label and value labels. If vallabs is NULL, the function returns x with a variable label stored as the attribute "label". If vallabs is not NULL, the function returns x of type haven::labelled with with the value labels stored in the attribute "labels" (For more information type ?haven::labelled in the console).

Examples

# Create toy dataframe:
df <-
  tibble::tibble(
    fbnr = 1:10,
    sex = haven::labelled(c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                          label = "sex",
                          labels = c(MALES = 1, FEMALES = 2)),
    age = c(24, 23, 23, 41, 23, 39, 30, 18, 31, 48),
    marital = haven::labelled(
      c(1, 7, 2, 6, 4, 5, 3, 8, 4, 2),
      label = "marital status",
      labels = c(
        "single" = 1,
        "steady relationship" = 2,
        "living with partner" = 3,
        "married first time" = 4,
        "remarried" = 5,
        "separated" = 6,
        "divorced" = 7,
        "non-occuring label" = 9
      )
    )
  )
df
#> # A tibble: 10 × 4
#>     fbnr         sex   age                 marital
#>    <int>   <dbl+lbl> <dbl>               <dbl+lbl>
#>  1     1 2 [FEMALES]    24 1 [single]             
#>  2     2 1 [MALES]      23 7 [divorced]           
#>  3     3 2 [FEMALES]    23 2 [steady relationship]
#>  4     4 1 [MALES]      41 6 [separated]          
#>  5     5 1 [MALES]      23 4 [married first time] 
#>  6     6 2 [FEMALES]    39 5 [remarried]          
#>  7     7 2 [FEMALES]    30 3 [living with partner]
#>  8     8 1 [MALES]      18 8                      
#>  9     9 2 [FEMALES]    31 4 [married first time] 
#> 10    10 1 [MALES]      48 2 [steady relationship]
# store all labels in df_attr:
df_attrs <- df %>% tab_attrs()
#> Error: Result 2 must be a single string, not a character vector of length 3
df_attrs
#> Error in eval(expr, envir, enclos): object 'df_attrs' not found
# Remove all attributes from df:
df <- df %>% dplyr::mutate_all(as.vector)
df
#> # A tibble: 10 × 4
#>     fbnr   sex   age marital
#>    <int> <dbl> <dbl>   <dbl>
#>  1     1     2    24       1
#>  2     2     1    23       7
#>  3     3     2    23       2
#>  4     4     1    41       6
#>  5     5     1    23       4
#>  6     6     2    39       5
#>  7     7     2    30       3
#>  8     8     1    18       8
#>  9     9     2    31       4
#> 10    10     1    48       2
# Write back label information from df_attrs
df <-
  df %>%
  purrr::map2_dfc(.,
                  df_attrs %>% purrr::transpose(),
                  ~{set_labs(.x, .y$labels, .y$varlab)})
#> Error in purrr::transpose(.): object 'df_attrs' not found
df
#> # A tibble: 10 × 4
#>     fbnr   sex   age marital
#>    <int> <dbl> <dbl>   <dbl>
#>  1     1     2    24       1
#>  2     2     1    23       7
#>  3     3     2    23       2
#>  4     4     1    41       6
#>  5     5     1    23       4
#>  6     6     2    39       5
#>  7     7     2    30       3
#>  8     8     1    18       8
#>  9     9     2    31       4
#> 10    10     1    48       2
# Constructed in this way, the dataframe including the label information could be
# exported to SPSS or Stata via haven:
# haven::write_sav(df, "test.sav")
# haven::write_dta(df, "test.dta")