Tabulate a dataframe's variables' attributes

Tabulate a dataframe's variables' attributes

tab_attrs(df)

Arguments

df
Dataframe

Value

Dataframe with one row per variable in df, containing the following columns: "var", "varlab", "val", "vallab", "format.spss", "labels", "display_width" & "class". In contrast to tab_types() the class shows the original type as it is imported by haven. This means labelled numerics or characters belong to the class haven_labelled.

Examples

path <- system.file("examples", "iris.sav", package = "haven")
df <- haven::read_sav(path) %>%
  # add id column
  tibble::rownames_to_column("id")

df %>% tab_attrs()

#> Error: Result 6 must be a single string, not a character vector of length 3

# Only show variables with a variable label:
df %>% tab_attrs() %>% dplyr::filter(!sapply(varlab, is.null))

#> Error: Result 6 must be a single string, not a character vector of length 3