Tabulate a dataframe's variables' value counts

Tabulate a dataframe's variables' value counts

tab_cts(df, id = "id", include_ids = FALSE)

Arguments

df
Dataframe

id
Character string of the id variable

include_ids
Logical denoting whether a list column ids should be included in the results. The ids in each list show at which values of id the variable var contains the value val.

Value

Dataframe consisting of columns var, val & n containing a tabulation of the counts of variables' values.

Examples

df <-
  tibble::tibble(
    fbnr = 1:10,
    sex = haven::labelled(c(2, 1, 2, 1, 1, 2, 2, 1, 2, 1),
                          label = "sex",
                          labels = c(MALES = 1, FEMALES = 2)),
    age = c(24, 23, 23, 41, 23, 39, 30, 18, 31, 48),
    marital = haven::labelled(
      c(1, 7, 2, 6, 4, 5, 3, 8, 4, 2),
      label = "marital status",
      labels = c(
        "single" = 1,
        "steady relationship" = 2,
        "living with partner" = 3,
        "married first time" = 4,
        "remarried" = 5,
        "separated" = 6,
        "divorced" = 7,
        "widowed" = 8
      )
    )
  )
tab_cts(df, "fbnr")

#> # A tibble: 18 × 4
#>    var        nv cv        n
#>    <chr>   <int> <chr> <int>
#>  1 sex         1 NA        5
#>  2 sex         2 NA        5
#>  3 age        18 NA        1
#>  4 age        23 NA        3
#>  5 age        24 NA        1
#>  6 age        30 NA        1
#>  7 age        31 NA        1
#>  8 age        39 NA        1
#>  9 age        41 NA        1
#> 10 age        48 NA        1
#> 11 marital     1 NA        1
#> 12 marital     2 NA        2
#> 13 marital     3 NA        1
#> 14 marital     4 NA        2
#> 15 marital     5 NA        1
#> 16 marital     6 NA        1
#> 17 marital     7 NA        1
#> 18 marital     8 NA        1